+++ components/ui/index.ts
@@ -5,6 +5,9 @@
 import Button from "./Button";
 import Confirm from "./Confirm";
 import Drawer from "./Drawer";
+import Input from "./Input";
 import Modal from "./Modal";
+import Select from "./Select";
+import Textarea from "./Textarea";
 
-export { Button, Confirm, Drawer, Modal };
+export { Button, Confirm, Drawer, Input, Modal, Select, Textarea };